pyproject.toml
[tool.poetry]
name = "nonebug-saa"
version = "0.2.0"
description = "A nonebug helper for nonebot-plugin-send-anything-anything"
authors = ["felinae98 <me@felinae98.cn>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/felinae98/nonebug-saa"
packages = [{include = "nonebug_saa"}]
repository = ""
documentation = ""
keywords = []
classifiers = ["Framework :: Pytest"]

[tool.poetry.dependencies]
python = "^3.10"
nonebug = "^0.3.1"
pytest-mock = "^3.10.0"
nonebot-plugin-send-anything-anywhere = "^0.2.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.3.0"
nonemoji = "^0.1.2"
pre-commit = "^2.16.0"

[tool.poetry.plugins.pytest11]
saa_nonebug = "nonebug_saa.fixtures"

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
